backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
Pillow>=10.0
